LayoutTests/imported/w3c/web-platform-tests/custom-elements/form-associated/ElementInternals-form-expected.txt


PASS ElementInternals.form should return the target element's form owner
PASS ElementInternals.form should throws NotSupportedError if the target element is not a form-associated custom element
